requirements.txt
Django>=4.0
numpy==1.17.3
django-crispy-forms==1.14.0
gunicorn==20.1.0
scipy==1.3.1
PyPDF2==1.27.2
Pillow==9.1.0
reportlab==3.6.9
pydub==0.25.1
mysql-connector-python==8.0.28

